# Reduced the number of times a lock is acquired.