--- a/hystrix/rolling.go
+++ b/hystrix/rolling.go
@@ -23,50 +23,43 @@ func newRollingNumber() *rollingNumber {
 }
 
 func (r *rollingNumber) getCurrentBucket() *numberBucket {
-	r.Mutex.RLock()
+	now := time.Now().Unix()
+	var bucket *numberBucket
+	var ok bool
 
-	now := time.Now()
-	bucket, exists := r.Buckets[now.Unix()]
-	if !exists {
-		r.Mutex.RUnlock()
-		r.Mutex.Lock()
-		defer r.Mutex.Unlock()
-
-		r.Buckets[now.Unix()] = &numberBucket{}
-		bucket = r.Buckets[now.Unix()]
-	} else {
-		defer r.Mutex.RUnlock()
+	if bucket, ok = r.Buckets[now]; !ok {
+		bucket = &numberBucket{}
+		r.Buckets[now] = bucket
 	}
+
 	return bucket
 }
 
 func (r *rollingNumber) removeOldBuckets() {
-	now := time.Now()
+	now := time.Now().Unix() - 10
 
 	for timestamp := range r.Buckets {
 		// TODO: configurable rolling window
-		if timestamp <= now.Unix()-10 {
+		if timestamp <= now {
 			delete(r.Buckets, timestamp)
 		}
 	}
 }
 
 func (r *rollingNumber) Increment() {
-	b := r.getCurrentBucket()
-
 	r.Mutex.Lock()
 	defer r.Mutex.Unlock()
 
+	b := r.getCurrentBucket()
 	b.Value++
 	r.removeOldBuckets()
 }
 
 func (r *rollingNumber) UpdateMax(n int) {
-	b := r.getCurrentBucket()
-
 	r.Mutex.Lock()
 	defer r.Mutex.Unlock()
 
+	b := r.getCurrentBucket()
 	if uint64(n) > b.Value {
 		b.Value = uint64(n)
 	}