There was a race condition in getCurrentBucket which is removed.  The codepaths that went down the getCurrentBucket path would sometimes acquire the same lock as many as 3 times.  Given that increment and set max would eventually have to acquire an exclusive lock, I just acquire that lock off the bat.  Added some benchmarks to state the case:
Current master:
BenchmarkRollingNumberIncrement-8	10000000	       204 ns/op
BenchmarkRollingNumberUpdateMax-8	10000000	       209 ns/op

This branch:
BenchmarkRollingNumberIncrement-8	10000000	       147 ns/op
BenchmarkRollingNumberUpdateMax-8	20000000	       149 ns/op

<!---
@huboard:{"order":29.0,"milestone_order":27,"custom_state":"blocked"}
-->