@afex 2015-04-16
thanks for the patch. i just merged a PR which refactored the rolling implementations.

could you sync your change with latest master?